Variational Gibbs with QPRC
Calculates the Gibbs Free Energy with the same method as the original file, but includes a method for Quasi-Probabilistic Readout Correction of Mid-Circuit Measurements which also makes use of dynamic circuits. It does not run in the state that it is uploaded because Qiskit Sampler and Estimator methods don't work with dynamic circuits. This can be used as a proof of concept and can be worked on to create a functional calculation of the Gibbs Free Energy with error correction.