Variational Quantum Gibbs with QPRC
Calculates the Gibbs Free Energy with the same method as the original file, but includes a method for Quasi-Probabilistic Readout Correction of Mid-Circuit Measurements which also makes use of dynamic circuits. Functioning with qiskit_ibm_runtime V2 primitives Sampler and Estimator, as well as Aer Simulator that can use noise models to simulate real life quantum hardware. The execution is very slow but arrives at a Free Energy that is similar to the Variational Gibbs calculation without QPRC.